requirements.txt
pandas
requests
pymongo
psycopg2-binary
fastapi
uvicorn
pytest
